fix nxos_vrf and migrate get_interface_type to module_utils
