+++ lib/ansible/modules/network/nxos/nxos_igmp_interface.py
@@ -191,6 +191,7 @@ changed:
 
 from ansible.module_utils.network.nxos.nxos import get_config, load_config, run_commands
 from ansible.module_utils.network.nxos.nxos import nxos_argument_spec, check_args
+from ansible.module_utils.network.nxos.nxos import get_interface_type
 from ansible.module_utils.basic import AnsibleModule
 
 import re
@@ -225,23 +226,6 @@ def get_interface_mode(interface, intf_type, module):
     elif intf_type == 'loopback' or intf_type == 'svi':
         mode = 'layer3'
     return mode
-
-
-def get_interface_type(interface):
-    if interface.upper().startswith('ET'):
-        return 'ethernet'
-    elif interface.upper().startswith('VL'):
-        return 'svi'
-    elif interface.upper().startswith('LO'):
-        return 'loopback'
-    elif interface.upper().startswith('MG'):
-        return 'management'
-    elif interface.upper().startswith('MA'):
-        return 'management'
-    elif interface.upper().startswith('PO'):
-        return 'portchannel'
-    else:
-        return 'unknown'
 
 
 def apply_key_map(key_map, table):
