+++ lib/ansible/modules/network/nxos/nxos_l2_interface.py
@@ -120,34 +120,9 @@ import re
 from copy import deepcopy
 
 from ansible.module_utils.network.nxos.nxos import get_config, load_config, run_commands
-from ansible.module_utils.network.nxos.nxos import nxos_argument_spec
+from ansible.module_utils.network.nxos.nxos import nxos_argument_spec, get_interface_type
 from ansible.module_utils.basic import AnsibleModule
 from ansible.module_utils.network.common.utils import remove_default_spec
-
-
-def get_interface_type(name):
-    """Gets the type of interface
-    Args:
-        interface (str): full name of interface, i.e. Ethernet1/1, loopback10,
-            port-channel20, vlan20
-    Returns:
-        type of interface: ethernet, svi, loopback, management, portchannel,
-         or unknown
-    """
-    if name.upper().startswith('ET'):
-        return 'ethernet'
-    elif name.upper().startswith('VL'):
-        return 'svi'
-    elif name.upper().startswith('LO'):
-        return 'loopback'
-    elif name.upper().startswith('MG'):
-        return 'management'
-    elif name.upper().startswith('MA'):
-        return 'management'
-    elif name.upper().startswith('PO'):
-        return 'portchannel'
-    else:
-        return 'unknown'
 
 
 def get_interface_mode(name, module):
